vmm: test programmatic API in rust integ tests
This commit demonstrates how to port a unit test for a pub fn
in vmm to a rustacean integration test for the vmm.
